Crate re_arrow_util

Helpers for working with arrow

Modules

arrays 🔒

batches 🔒

compare 🔒

format_data_type 🔒

arrow has ToString implemented, but it is way too verbose.

Traits

ArrowArrayDowncastRef

Downcast an arrow array to another array, without having to go via Any.

Functions

almost_equal_f64

Return true when arguments are the same within some rounding error.

arrays_to_list_array

Create a sparse list-array out of an array of arrays.

arrays_to_list_array_opt

Create a sparse list-array out of an array of arrays.

concat_arrays

Applies a [arrow::compute::concat] kernel to the given arrays.

concat_polymorphic_batches

Concatenates the given [RecordBatch]es, regardless of their respective schema.

ensure_similar

Are two arrays equal, ignoring small numeric differences?

extract_fixed_size_array_element

Extract the element at idx from a FixedSizeListArray.

filter_array

Applies a filter kernel to the given array.

format_data_type

Compact format of an arrow data type.

into_arrow_ref

is_list_array_semantically_empty

Returns true if the given list_array is semantically empty.

make_batch_nullable

Returns a new [RecordBatch] where all top-level fields are nullable.

new_list_array_of_empties

Returns a new [[ListArray]] with len entries.

offsets_lengths

Returns an iterator with the lengths of the offsets.

pad_list_array_back

Create a new [ListArray] of target length by appending null values to its back.

pad_list_array_front

Create a new [ListArray] of target length by appending null values to its front.

repartition_list_array

Repartitions a [ListArray] according to the specified lengths, ignoring previous partitioning.

sparse_list_array_to_dense_list_array

Given a sparse [ListArray] (i.e. an array with a nulls bitmap that contains at least one falsy value), returns a dense [ListArray] that only contains the non-null values from the original list.

take_array

Applies a take kernel to the given array.

wrap_in_list_array

Convert [A, B, null, D, …] into [[A], [B], null, [D], …].