re_arrow_util::arrays
Trait ArrowArrayDowncastRef

pub trait ArrowArrayDowncastRef<'a>: 'a {
    // Required methods
    fn downcast_array_ref<T: Array + 'static>(self) -> Option<&'a T>;
    fn try_downcast_array_ref<T: Array + 'static>(
        self,
    ) -> Result<&'a T, ArrowError>;
}

Downcast an arrow array to another array, without having to go via Any.

Required Methods

fn downcast_array_ref<T: Array + 'static>(self) -> Option<&'a T>

Downcast an arrow array to another array, without having to go via Any.

fn try_downcast_array_ref<T: Array + 'static>(self) -> Result<&'a T, ArrowError>

Similar to downcast_array_ref, but returns an error in case the downcast returns None.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.