re_arrow_util::batches
Function concat_polymorphic_batches

pub fn concat_polymorphic_batches(
    batches: &[RecordBatch],
) -> Result<RecordBatch>

Concatenates the given [RecordBatch]es, regardless of their respective schema.

The final schema will be the merge of all the input schemas.

This will fail if the concatenation requires backfilling null values into non-nullable column. You probably want to call make_batch_nullable first.