re_arrow_util
Function take_array

pub fn take_array<A, O>(array: &A, indices: &PrimitiveArray<O>) -> A
where
    A: Array + Clone + 'static,
    O: ArrowPrimitiveType,
    O::Native: Add<Output = O::Native>,

Applies a take kernel to the given array.

In release builds, indices are allowed to have null entries (they will be taken as nulls). In debug builds, null entries will panic.

Note: a take kernel copies the data in order to make the resulting arrays contiguous in memory.

Takes care of up- and down-casting the data back and forth on behalf of the caller.